remove source files that do not implement any symbols declared in the header file (test ok for pure c)

+++ test_sources/test_c/src/test_functions.h
@@ -4,7 +4,7 @@
 
 // normal callbacks
 typedef void (*simple_callback)(void);
-typedef int (*data_callback)(const void *, size_t);
+typedef int (*data_callback)(const void *, int);
 
 // complex callbacks
 struct CallbackInfo
